The 'RTestCase' Class

Objects of the RTestCase class specify and implement the general behavoir of the XML-based test cases (TC).

Details

On basis of Based on a XML definition file (see 'RTestCase.xsd'), test cases are defined for the individual packages / projects to be tested. Thereby, this class definition provides the general outline of the individual test case classes (i.e. test case adapters), which have to implement the project- and package-specific requirements and test logic.

The slot 'tests' represents the execution environment for a test case and stores all information about the execution including test case metainformation, test result (success or failed), the testthat reporter and the test cache. It has the following format:
XML: tests:
tests
|- Pkg_1 [Pkg_1][1]
| - funct_01 [Pkg_1][1][funct_01][1] = list( ... )
|- Pkg_1 [Pkg_1][2]
| |- funct_01 [Pkg_1][2][funct_01][1] = list( ... )
| |- funct_01 [Pkg_1][2][funct_01][2] = list( ... )
| - funct_02 [Pkg_1][2][funct_02][1] = list( ... )
- Pkg_2 [Pkg_2][1]
- funct_01 [Pkg_2][1][funct_01][1] = list( ... )

Slots

ID

(character) ID of the TC.

tc.type

(character) Type of the TC (i.e. the class name).

synopsis

(list) Synopsis information of the TC (as defined in the XSD ComplexType 'RTestSynopis').

xml.fPath

(character) Path to XML definition file of the TC.

xml.root

(XMLNode) The imported TC definition as XMLNode-class object.

input.data

(list) The input data of the test case, which will be filled using the method readXMLInputData.

tests

(list) A list with the test results or NULL, if the test case has not been executed (see 'Details').

test.for

(character) Names of the packages, which were tested in the last execution of the test case. NULL, if the test case has not been executed.

test.result

(character) Result of the last test case execution ('SUCCESS' or 'FAILURE'). NULL, if the test case has not been executed.