Read arguments from RTest 'param'-XML Node

Read arguments from RTest 'param'-XML Node

arguments_creator(parameters_xml_definition, input_data = NULL)

Arguments

parameters_xml_definition
(XMLNode) An XML Node that has elements of kind RTestData_variable, RTestData_image, RTestData_vector, RTestData_list, RTestData_data.frame in it. Please see the RTest.xsd to be found in file.path(find.package("RTest")/"xsd/RTest.xsd")

input_data
(list) A named list of values of kind data.frame, character, numeric or list of those

Value

A named list. For the name of the RTestData_input_data element, please see details.

Details

This function will read in all parameters exept the one named "RTestData_input_data" into a list by using xmlReadData_to_list. The parameter "RTestData_input_data" is written into an additional item of the list. The name of this item is given by the "param" attribute of the XMLNode "RTestData_input_data". The value is given by the list item of the list "input_data" that can be found by the "name" attribute of the XMLNode "RTestData_input_data".