Show Summary of RTestCollection Instance

Show Summary of RTestCollection Instance

# S4 method for RTestCollection
show(object)

Arguments

object
(object) The RTestCollection-class object.

See also

Examples

location <- find.package("RTest")

testCollection <- new("RTestCollection",
   project.name    = "RTest Vignette",
   project.details = "Example test exectuion",
   tester          = "Example tester",
   test.start      = format(Sys.time(), "%Y-%m-%d %H:%M:%S"))

show(testCollection)

#> RTestCollection
#>   @project.name:       RTest Vignette 
#>   @project.details:    Example test exectuion 
#>   @tester:             Example tester 
#>   @test.start:         2019-12-04 15:31:11 
#>   @collection:         0  TCs
#> 

testCollection <- importTCsFromDir(testCollection,
   xml.dPath = paste0(location,"/xml-templates")
)

#> 10 files found in input directory:
#>  [1] "RTest_medium.xml"     "RTest_TC-01.xml"      "RTest_TC-02.xml"     
#>  [4] "RTest_TC-fail.xml"    "RTest_TC-generic.xml" "test.execCache.xml"  
#>  [7] "test.execCache2.xml"  "test.large.xml"       "test.RTM.xml"        
#> [10] "test.testthat.xml"   
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_medium.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-medium 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Sebastian Wolf 
#>      $short-description :   
#>      $description :   
#>      $label :   
#>      $creation-date :  2018-12-11 
#>      $change-history :
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_medium.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 1
#>   ..$ my_function:List of 2
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "my_function"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::my_function"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.my_function"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "Test data.frame"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/my_function[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "my_function"
#>   .. .. ..$ func-iter   : num 2
#>   .. .. ..$ test-code   : chr "RTest::my_function"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.my_function"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "Test data.frame"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/my_function[2]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-01.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-01 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :   
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-01.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 1
#>   ..$ funct_01:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_01"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_01"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_01"
#>   .. .. ..$ pkg-desc    : chr "First test for Pkg_1"
#>   .. .. ..$ func-desc   : chr "First test of funct_01"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_01[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-02.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-02 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Sebastian Wolf 
#>      $short-description :  RTest Template TC 
#>      $description :  A test case without a test adapter 
#>      $label :   
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-02.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 1
#>   ..$ funct_02:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_02"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_02"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_02"
#>   .. .. ..$ pkg-desc    : chr "First test for Pkg_1"
#>   .. .. ..$ func-desc   : chr "First test of funct_01"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_02[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-fail.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-fail 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :   
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-fail.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 2
#>  $ :List of 1
#>   ..$ funct_01:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_01"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_01"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_01"
#>   .. .. ..$ pkg-desc    : chr "First test for Pkg_1"
#>   .. .. ..$ func-desc   : chr "First test of funct_01"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_01[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#>  $ :List of 1
#>   ..$ funct_01:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 2
#>   .. .. ..$ func        : chr "funct_01"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_01"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_01"
#>   .. .. ..$ pkg-desc    : chr "Failing test"
#>   .. .. ..$ func-desc   : chr "see test_fun fail"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[2]/funct_01[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-generic.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-generic 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Sebastian Wolf 
#>      $short-description :  RTest Template TC for generic test-adapter 
#>      $description :  A test case without a test adapter 
#>      $label :   
#>      $creation-date :  2018-12-11 
#>      $change-history :
#>         2018-12-11 ,  Sebastian Wolf :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/RTest_TC-generic.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 5
#>  $ :List of 1
#>   ..$ example_data_frame:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "example_data_frame"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::example_data_frame"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.example_data_frame"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "Test data.frame"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/example_data_frame[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "01"
#>   .. .. ..$ riskid      : chr "01"
#>  $ :List of 1
#>   ..$ example_image:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 2
#>   .. .. ..$ func        : chr "example_image"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::example_image"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.example_image"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "Compare an image"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[2]/example_image[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "02"
#>   .. .. ..$ riskid      : chr "01"
#>  $ :List of 1
#>   ..$ example_list:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 3
#>   .. .. ..$ func        : chr "example_list"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::example_list"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.example_list"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[3]/example_list[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "03"
#>   .. .. ..$ riskid      : chr "01"
#>  $ :List of 1
#>   ..$ example_vector:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 4
#>   .. .. ..$ func        : chr "example_vector"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::example_vector"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.example_vector"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "Test vector example"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[4]/example_vector[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "04"
#>   .. .. ..$ riskid      : chr "01"
#>  $ :List of 1
#>   ..$ example_variable:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 5
#>   .. .. ..$ func        : chr "example_variable"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::example_variable"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.example_variable"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "Test relative difference"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[5]/example_variable[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "05"
#>   .. .. ..$ riskid      : chr "01"
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.execCache.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-04 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :  My Label 
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.execCache.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 3
#>   ..$ funct_01:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_01"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_01"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_01"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "see test_fun fail"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_01[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "2"
#>   .. .. ..$ riskid      : chr "1"
#>   ..$ funct_02:List of 2
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_02"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_02"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_02"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_02[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "3"
#>   .. .. ..$ riskid      : chr "1"
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_02"
#>   .. .. ..$ func-iter   : num 2
#>   .. .. ..$ test-code   : chr "RTest::funct_02"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_02"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_02[2]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#>   ..$ funct_03:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_03"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_03"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_03"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_03[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.execCache2.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-cache2 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :  My Label 
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.execCache2.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 1
#>   ..$ funct_04:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_04"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_04"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_04"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_04[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.large.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-Large 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :  My Label 
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.large.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 1
#>   ..$ example_list_large:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "example_list_large"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::example_list_large"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.example_list_large"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/example_list_large[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.RTM.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-03 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :  My Label 
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.RTM.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 3
#>   ..$ funct_01:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_01"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_01"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_01"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr "see test_fun fail"
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_01[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "2"
#>   .. .. ..$ riskid      : chr "1"
#>   ..$ funct_02:List of 2
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_02"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_02"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_02"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_02[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "3"
#>   .. .. ..$ riskid      : chr "1"
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_02"
#>   .. .. ..$ func-iter   : num 2
#>   .. .. ..$ test-code   : chr "RTest::funct_02"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_02"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_02[2]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#>   ..$ funct_03:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "funct_03"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::funct_03"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.funct_03"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/funct_03[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> secondpackage : List of 1
#>  $ :List of 1
#>   ..$ fun_01:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "secondpackage"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "fun_01"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "secondpackage::fun_01"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.secondpackage.fun_01"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/secondpackage[1]/fun_01[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr "1"
#>   .. .. ..$ riskid      : chr "2"
#> 
#> Import TC file '/tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.testthat.xml'
#> 
#> Create new TC using adapter definition 'RTestCase'.Test Case
#>   Object of class ' RTestCase '
#>   @ID :  RTest_TC-testthat 
#>   @tc.type :  RTestCase 
#>   @synopsis :
#>      $version :  01 
#>      $author :  Matthias Pfeifer 
#>      $short-description :  RTest Template TC 
#>      $description :  Extended Description of the test case allowing also <some><special>/characters 
#>      $label :  My Label 
#>      $creation-date :  2016-01-25 
#>      $change-history :
#>         2016-01-25 ,  Matthias Pfeifer :
#>           Initial Version 
#> 
#>   @xml.fPath :  /tmp/RtmpSMomZh/temp_libpath6c0625f3fa09/RTest/xml-templates/test.testthat.xml 
#>   @xml.root :  XMLNode 
#> 
#>   @input.data : 
#> test01 : 'data.frame':	2 obs. of  2 variables:
#>  $ x: num  1 1
#>  $ y: num  2 2
#> 
#>   @test.for :  NA 
#>   @test.result :  NA 
#> 
#>   @tests :
#> RTest : List of 1
#>  $ :List of 1
#>   ..$ my_fun:List of 1
#>   .. ..$ :List of 16
#>   .. .. ..$ pkg         : chr "RTest"
#>   .. .. ..$ pgk-iter    : num 1
#>   .. .. ..$ func        : chr "my_fun"
#>   .. .. ..$ func-iter   : num 1
#>   .. .. ..$ test-code   : chr "RTest::my_fun"
#>   .. .. ..$ test-adapter: chr "RTestCase"
#>   .. .. ..$ test-func   : chr "test.RTest.my_fun"
#>   .. .. ..$ pkg-desc    : chr ""
#>   .. .. ..$ func-desc   : chr ""
#>   .. .. ..$ xpath       : chr "//RTestCase/tests/RTest[1]/my_fun[1]"
#>   .. .. ..$ reporter    : logi NA
#>   .. .. ..$ result      : logi NA
#>   .. .. ..$ cache       : logi NA
#>   .. .. ..$ execresid   : logi NA
#>   .. .. ..$ specid      : chr ""
#>   .. .. ..$ riskid      : chr ""
#> 
#> 

# Now one test case shall be imported
show(testCollection)

#> RTestCollection
#>   @project.name:       RTest Vignette 
#>   @project.details:    Example test exectuion 
#>   @tester:             Example tester 
#>   @test.start:         2019-12-04 15:31:11 
#>   @collection:         10  TCs
#>